Separated State into multiple files and refactored the general logic.
With multiple files, not every state can import every other state because of circular imports. Therefore, a state machine is used to prevent this problem.
