+++ hand-painter/FreeModeState.py
@@ -1,0 +1,17 @@
+from cv2 import Mat
+
+from PaintingState import PaintingState
+from State import State
+from Hand import Hand
+
+class FreeModeState(PaintingState):
+    def run(self, img, hands: list[Hand]) -> tuple["State", Mat]:
+        self.paint(img, hands)
+        state, img = self.draw_menu(img, hands)
+        img = self.picture_btn.draw(img, hands)
+
+        for hand in hands:
+            if self.picture_btn.click(hand):
+                return self.stateMachine.pictureTimerState(), img
+
+        return state, img
